backend/requirements.txt
bcrypt==4.0.1
firebase-admin==6.0.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.1.1
gunicorn==20.1.0
numpy>=1.23,<1.26
pandas==2.1.1
python-dotenv==1.0.0
yfinance==0.2.28
